# specs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Checkmate API
  version: 1.0.0
  summary: Self-hosted personal task manager.
  description: |
    A single-user task manager organised around four parallel **contexts**
    (Upsun, Personal, Gaal, Arkea): capture without choosing a context, triage
    deliberately, and read one daily brief.

    ## Authentication

    Three credential kinds, all resolving to the same user:

    | Kind | Used by | Obtained |
    |---|---|---|
    | Session cookie | Web | Google OIDC via `/auth/login/google` |
    | Device token (`cm_…`) | iOS, macOS, CLI | `POST /v1/tokens` (session only) or the CLI |
    | OAuth access token (`cmat_…`) | Remote MCP clients | `/oauth/authorize` + `/oauth/token` |

    Scopes are `read` (all GETs) and `write` (all mutations). A session cookie
    always carries both.

    ## Conventions

    - Collections return `{"data": [...], "next_cursor": "<id>"|null}`; single
      resources return the bare object.
    - Ids are UUIDv7 strings. Because v7 sorts chronologically, `id DESC` is
      creation order, which is how every list is ordered.
    - Timestamps are RFC3339 UTC (`2026-07-25T14:03:11.482Z`). Calendar dates are
      plain `YYYY-MM-DD` with no timezone.
    - Estimates are whole minutes.
    - Errors are `{"error": "...", "fields": {"due_on": "..."}}`. When `fields` is
      present, render each entry against its control rather than as a toast.

    ## Ordering

...
    opens with undated tasks.

...
  contact:
    name: Checkmate
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /auth/login/{provider}:
    get:
      tags: [Authentication]
      summary: Begin federated sign-in
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, examples: ["google"] }
        - name: redirect_to
          in: query
          description: |
            Where to land after sign-in. Reduced to a same-site path; an absolute
            URL, a scheme-relative `//host` or anything not rooted at `/` is
            discarded in favour of `/`.
          schema: { type: string, examples: ["/tasks/abc123"] }
      responses:
        "302": { description: Redirect to the identity provider. }
        "400": { description: Unknown provider. }
        "501": { description: No identity provider is configured. }
  /auth/callback/{provider}:
    get:
      tags: [Authentication]
      summary: Finish federated sign-in
      description: |
        Sets the session cookie and redirects. Single-use: the login flow is
        consumed on arrival, so a replayed callback URL fails.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: code
          in: query
          schema: { type: string }
        - name: error
          in: query
          description: Set when the provider declined.
          schema: { type: string }
      responses:
        "302": { description: Signed in; cookie set. }
        "400":
          description: |
            Missing state or code, or the link has expired or was already used.
        "403":
          description: |
            The address is not on `CHECKMATE_ALLOWED_EMAILS`, or the provider did
            not report the email as verified. Not recoverable by retrying, and
            must not be presented as a wrong-password error.
        "501": { description: No identity provider is configured. }
  /v1/me:
    get:
      tags: [Authentication]
      summary: The authenticated user
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    patch:
      tags: [Authentication]
      summary: Update your own profile
      description: |
        `name` and `timezone` only. **Email is not editable here** — it is the join
        key for a federated identity, so changing it would move which provider
        account owns the data, and belongs in its own deliberate flow.

        Timezone is why this exists: it decides which day "today" is for the brief,
        and an account provisioned with the wrong default previously had no way to
        correct it from any client.

        The `users` table carries no `rev`, so a change here does **not** appear in
        the sync feed. Re-read this endpoint on foreground rather than caching the
        profile indefinitely.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, minLength: 1 }
                timezone:
                  type: string
                  description: IANA name. Not nullable.
                  examples: ["Europe/Paris"]
      responses:
        "200":
          description: Updated, in the same shape as GET.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/logout:
    post:
      tags: [Authentication]
      summary: End the current session
      parameters:
        - name: everywhere
          in: query
          description: Revoke every session for this user, not just this one.
          schema: { type: boolean, default: false }
      responses:
        "204": { description: Signed out; cookie cleared. }
        "400":
          description: |
            The caller authenticated with a bearer token, which has no session to
            end. Revoke the token instead.
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/sources:
    get:
      tags: [Tasks]
      summary: The fixed source lookup
      description: |
        Six rows, global rather than per-user, and effectively immutable: `self`,
        `email`, `slack`, `google_chat`, `meeting`, `phone`.
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Source" }
                  next_cursor: { type: ["string", "null"] }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/tasks/{id}/restore:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      tags: [Tasks]
      summary: Undo a task deletion
      description: |
        Deletes are tombstones, so the data is still there; this is the only way
        back. Restores the task and **exactly the subtree that went down with it**,
        identified by the batch id that delete stamped — a child deleted separately
        and earlier stays deleted.

        References that died while the task was away are repaired rather than
        restored into an incoherent row: a deleted context or project is dropped
        (returning the task to the inbox), a still-deleted parent is dropped
        (promoting it to top level), and a deleted delegate is dropped with the
        status returning to `todo`.

        **What restore cannot do** is re-establish edges that pointed *at* this
        task: the deletion cleared `blocked_by_id` on its dependents and that
        information is gone.

        Tasks only. Contexts, projects, people and recurrences have no restore, so
        their delete confirmations must be firm.
      responses:
        "200":
          description: Restored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: |
            The task is not deleted. Reported as a validation problem on `id`
            rather than a conflict, so an idempotent retry does not look broken.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/people/{id}/merge:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      tags: [People]
      summary: Fold one person into another
      description: |
        Delegating by name creates a person on the spot, which is the right
        ergonomics for fast capture and guarantees that "Marc", "marc" and
        "Marc D." accumulate. This is the way back.

        Every task and recurrence pointing at `{id}` is repointed at `into`, and
        `{id}` is tombstoned. The target keeps its own name, email and notes. A
        delegated task never has a moment without a delegate, so nothing has to be
        un-delegated on the way through.

        Not reversible beyond re-delegating by hand, so confirm before calling.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [into]
              additionalProperties: false
              properties:
                into:
                  type: string
                  description: The person to keep. Must differ from `{id}`.
      responses:
        "200":
          description: Merged.
          content:
            application/json:
              schema:
                type: object
                required: [person, tasks_moved, merged_id]
                properties:
                  person: { $ref: "#/components/schemas/Person" }
                  tasks_moved:
                    type: integer
                    description: |
                      How many tasks were repointed, so a client can say "3 tasks
                      moved to Marc" rather than reporting a silent success.
                  merged_id:
                    type: string
                    description: The person that was tombstoned.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: |
            Validation failed. `into` when it is missing, equal to `{id}`, or names
            a person who does not exist.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/brief:
    get:
      tags: [Brief]
      summary: The daily brief
      description: |
        Eight buckets plus totals, each bucket sorted server-side. Buckets
        overlap: one task can be in `overdue`, `planned` and `blocked` at once.

        Lists are capped at 100 entries; `totals` are exact and uncapped, so
        `totals.overdue` may exceed `overdue.length`.

        `inbox` ignores `context_id`, because an untriaged task has no context and
        filtering would always empty it. Label the count as global when a context
        filter is active.
      parameters:
        - name: date
          in: query
          description: Defaults to today in the caller's account timezone.
          schema: { $ref: "#/components/schemas/Date" }
        - name: timezone
          in: query
          description: |
            IANA name overriding the account timezone, deciding which day "today"
            is. Echoed back in the response.
          schema: { type: string, examples: ["Europe/Paris"] }
        - name: context_id
          in: query
          description: Limits every bucket except `inbox`.
          schema: { type: string }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Brief" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422":
          description: |
            Validation failed. `timezone` for an unknown IANA name, `date` for a
            non-existent calendar date such as `2026-02-31`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/sync:
    get:
      tags: [Sync]
      summary: Changes since a cursor
      description: |
        A one-way feed: the server is authoritative and there is no offline write
        path. Persist `cursor`, replay the rows, and apply any row carrying
        `deleted_at` as a deletion.

        `sources` is present only on a full sync (`since=0`), since it has no
        `rev`. Every collection is always an array, never null. `has_more` means
        more changes are already waiting, so poll again immediately rather than
        waiting for the next tick.
      parameters:
        - name: since
          in: query
          description: The last `cursor` received. Omit or `0` for a full sync.
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: limit
          in: query
          description: Rows per table.
          schema: { type: integer, minimum: 1, maximum: 2000, default: 500 }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/tokens:
    get:
      tags: [Tokens]
      summary: List device tokens
      description: The secret is not stored, so it never appears here.
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Tokens]
      summary: Issue a device token
      description: |
        **Session cookie only.** A bearer-token caller gets 403: minting a
        long-lived credential from a long-lived credential would let a leaked
        token renew itself indefinitely.

        The secret is returned **once** and only its hash is kept.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreate" }
      responses:
        "201":
          description: Created. `token` is the only copy of the secret.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: |
            The caller used a bearer token, or lacks the `write` scope.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/tokens/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    delete:
      tags: [Tokens]
      summary: Revoke a device token
      description: Takes effect immediately. Revoking twice is not an error.
      responses:
        "204": { description: Revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/grants:
    get:
      tags: [OAuth]
      summary: List connected OAuth clients
      description: One entry per client the user has consented to.
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GrantList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/grants/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    delete:
      tags: [OAuth]
      summary: Withdraw consent
      description: |
        Revokes every access and refresh token issued under the grant and deletes
        its outstanding authorization codes. Immediate.
      responses:
        "204": { description: Withdrawn. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /.well-known/oauth-authorization-server:
    get:
      tags: [OAuth]
      summary: Authorization server metadata (RFC 8414)
      security: []
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                description: |
                  Includes `code_challenge_methods_supported: ["S256"]`, which MCP
                  clients treat as proof that PKCE is supported;
                  `authorization_response_iss_parameter_supported: true` (RFC
                  9207); and `client_id_metadata_document_supported: true`.
                additionalProperties: true
        "404": { description: The authorization server is disabled. }
  /.well-known/oauth-protected-resource:
    get:
      tags: [OAuth]
      summary: Protected resource metadata (RFC 9728)
      description: |
        Also served path-inserted at
        `/.well-known/oauth-protected-resource/mcp`, which is what a client
        probing for the MCP endpoint's metadata requests.
      security: []
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                type: object
                required: [resource, authorization_servers]
                properties:
                  resource: { type: string }
                  authorization_servers:
                    type: array
                    items: { type: string }
                  scopes_supported:
                    type: array
                    items: { type: string }
                  bearer_methods_supported:
                    type: array
                    items: { type: string, examples: ["header"] }
                  resource_name: { type: string }
        "404": { description: Unknown resource, or the authorization server is disabled. }
  /oauth/authorize:
    get:
      tags: [OAuth]
      summary: Authorization endpoint (consent screen)
      description: |
        Renders HTML, not JSON. An unauthenticated visitor is redirected into
        sign-in and returned here with the request intact.

        PKCE is mandatory: `code_challenge` with `code_challenge_method=S256`. An
        omitted method is an error rather than RFC 7636's `plain` default.
        Redirect URIs are matched exactly. Errors are only delivered to the
        redirect URI once it has been verified against the client; before that
        they render as a page, so this endpoint cannot be used as an open redirect.
      security: []
      parameters:
        - { name: client_id, in: query, required: true, schema: { type: string } }
        - { name: redirect_uri, in: query, schema: { type: string } }
        - { name: response_type, in: query, required: true, schema: { type: string, const: code } }
        - { name: scope, in: query, schema: { type: string, examples: ["read write offline_access"] } }
        - { name: state, in: query, schema: { type: string } }
        - { name: code_challenge, in: query, required: true, schema: { type: string } }
        - { name: code_challenge_method, in: query, required: true, schema: { type: string, const: S256 } }
        - name: resource
          in: query
          description: RFC 8707 resource indicator. Defaults to this server.
          schema: { type: string }
      responses:
        "200": { description: The consent screen, as HTML. }
        "302": { description: Redirect to sign-in, or an error response to a verified redirect URI. }
        "400": { description: Malformed request, rendered as HTML. }
        "401": { description: Unknown client, rendered as HTML. }
        "501": { description: No identity provider, so consent cannot be obtained. }
    post:
      tags: [OAuth]
      summary: Record the consent decision
      description: |
        Form-encoded, session-authenticated, same-origin enforced. On approval,
        redirects to the client with `code`, `state` and `iss`.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [request_id, decision]
              properties:
                request_id: { type: string }
                decision: { type: string, enum: [approve, deny] }
      responses:
        "302": { description: Redirect to the client's redirect URI. }
        "400": { description: The request expired or was already used. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Cross-origin request refused. }
  /oauth/token:
    post:
      tags: [OAuth]
      summary: Token endpoint
      description: |
        Grants: `authorization_code` and `refresh_token`. Refresh tokens rotate on
        every use; replaying a rotated one revokes the whole grant, on the
        assumption that a replay means the token leaked.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type: { type: string, enum: [authorization_code, refresh_token] }
                code: { type: string }
                redirect_uri: { type: string }
                code_verifier: { type: string, minLength: 43, maxLength: 128 }
                refresh_token: { type: string }
                client_id: { type: string }
                client_secret: { type: string }
                resource: { type: string }
                scope:
                  type: string
                  description: On refresh, may narrow the grant but never widen it.
      responses:
        "200":
          description: Issued. Never cached.
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400":
          description: |
            `invalid_request`, `invalid_grant`, `invalid_scope`, `invalid_target`
            or `unsupported_grant_type`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthError" }
        "401":
          description: "`invalid_client`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthError" }
  /oauth/revoke:
    post:
      tags: [OAuth]
      summary: Revocation endpoint (RFC 7009)
      description: |
        Returns 200 whether or not the token existed, so it cannot be used to
        probe. Revoking a refresh token also revokes its grant's access tokens.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
                token_type_hint: { type: string, enum: [access_token, refresh_token] }
                client_id: { type: string }
      responses:
        "200": { description: Done. }
        "400":
          description: Malformed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthError" }
        "401":
          description: "`invalid_client`: no client_id."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthError" }
  /oauth/register:
    post:
      tags: [OAuth]
      summary: Dynamic client registration (RFC 7591)
      description: |
        Open registration, capped by `CHECKMATE_OAUTH_MAX_DYNAMIC_CLIENTS`.
        Deprecated by the MCP draft spec in favour of Client ID Metadata
        Documents, and disableable with `CHECKMATE_OAUTH_ALLOW_DCR=false`.

        Unknown metadata is ignored rather than rejected, as RFC 7591 expects.
        Redirect URIs must be https, or loopback http for a native client.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientRegistration" }
      responses:
        "201":
          description: |
            Registered. `client_secret` appears only for a confidential client.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClientRegistered" }
        "400":
          description: Invalid metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthError" }
        "403":
          description: Registration is disabled, or the client cap is reached.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OAuthError" }
        "404": { description: The authorization server is disabled. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A device token (`cm_…`) or an OAuth access token (`cmat_…`). An OAuth token
        is additionally checked to have been issued for this resource. A refresh
        token presented here is refused.
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-checkmate_session
      description: |
        Set by the sign-in callback. `HttpOnly`, `Secure`, `SameSite=Lax`. The
        `__Host-` prefix requires `Secure`, so a development server over plain
        http uses the unprefixed name `checkmate_session` instead.

        Cookie-authenticated mutations must additionally prove same origin, via
        `Sec-Fetch-Site` or a matching `Origin`; a request with neither is refused.
    oauth2:
      type: oauth2
      description: For remote MCP clients. PKCE with S256 is mandatory.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            read: See tasks, projects, contexts and people.
            write: Create, change and delete.
            offline_access: Receive a refresh token.
...
  responses:
...
    Unauthorized:
      description: |
        No credential, or one that is unknown, revoked, expired, or issued for a
        different resource. `WWW-Authenticate` names the protected resource
        metadata document and the scopes the request needs.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...